Add ownership-based authorization for alumni edit and delete

+++ modules/alumni/delete_alumni.php
@@ -5,7 +5,14 @@ include "../../config/database.php";
 
 $id = $_GET['id'];
 
-$query = "DELETE FROM alumni WHERE id='$id'";
+$user_id = $_SESSION['user_id'];
+
+$query = "DELETE FROM alumni WHERE id='$id' AND user_id='$user_id'";
+
+if (mysqli_affected_rows($conn) == 0) {
+    echo "Unauthorized or record not found";
+    exit();
+}
 
 if (mysqli_query($conn, $query)) {
     header("Location: alumni_list.php?deleted=1");
